Renamed some things.
I was trying to write some code, but I was too confused because all the method names seemed to be describing such simillar things. I tried to fix that so it makes more sense. Feel free to revert this commit if you dont like my changes. I will not get that mad at you.